requirements.txt
python-telegram-bot
yfinance
pandas
plotly
pytz
google-api-python-client
youtube-transcript-api
google-generativeai
twikit
pandas-market-calendars
python-dotenv